requirements.txt
torch
faiss-cpu
sentence-transformers
fastapi
uvicorn
pydantic
numpy
pandas
rich
